Move HistorySectionContent to a dedicated file
